requirements.txt
opencv-python==4.5.2.54
paddleseg
GDAL
